feat(zsh): shell updates 🐚

+++ zsh/.config/zsh/06-aliases.zsh
@@ -1,32 +1,39 @@
-#                 ██
-#  ██████  ██████░██
-# ░░░░██  ██░░░░ ░██████
-#    ██  ░░█████ ░██░░░██
-#   ██    ░░░░░██░██  ░██
-#  ██████ ██████ ░██  ░██
-# ░░░░░░ ░░░░░░  ░░   ░░
+# █▀▀██ ▄█▀█ ▄█ █ ▄█▀█ ▄█   ▄█
+#   ▄█▀ ▓█▄▄ ▓█ █ ▓█▄  ▓█   ▓█
+# ▄█▀   ▄ ██ ▓█▀█ ▓█ ▄ ▓█ ▄ ▓█ ▄
+# ▓█▄▄█ ▓▄█▀ ▓█ █ ▓█▄█ ▓█▄█ ▓█▄█
 #
-#  ▓▓▓▓▓▓▓▓▓▓
-# ░▓ author ▓ xero <x@xero.style>
-# ░▓ code   ▓ https://code.x-e.ro/dotfiles
-# ░▓ mirror ▓ https://git.io/.files
-# ░▓▓▓▓▓▓▓▓▓▓
-# ░░░░░░░░░░
-#
+# ░ config from xero's dotfiles
+# ▒ author: xero (x@xero.style)
+# ▓ https://git.io/.files
+# █ https://code.x-e.ro/dotfiles
+
 #█▓▒░ aliases
 alias c="clear"
 alias ll="ls -lahF --color=always"
 alias e="$EDITOR"
 alias se="sudoedit"
-alias ec='nvim --cmd ":lua vim.g.noplugins=1" ' #nvim --clean
+alias ec='nvim --cmd ":lua vim.g.noplugins=1"'
 alias fuck='sudo $(fc -ln -1)'
 alias g="git"
 alias y="yank"
-alias k="vpnns -- kubectl"
+alias k="kubectl"
 alias kx="kubectx"
-alias k9s="vpnns -- k9s"
-alias tgp="vpnns -- terragrunt plan"
-alias tga="vpnns -- terragrunt apply"
+alias tgp="terragrunt plan"
+alias tga="terragrunt apply"
+alias syncmusic="rsync -av --progress \
+  --include='*/' \
+  --include='*.mp3' \
+  --include='*.flac' \
+  --include='*.m4a' \
+  --include='*.wav' \
+  --include='*.ogg' \
+  --include='*.aac' \
+  --include='*.wma' \
+  --include='*.alac' \
+  --include='*.aiff' \
+  --exclude='.*' \
+  --exclude='*'"
 
 #lazy
 alias "cd.."="cd ../"
@@ -35,26 +42,28 @@ alias psef="ps -ef"
 alias ZZ="exit"
 
 #git
-alias ga="git add"
-alias gb="git branch"
-alias gc="git clone"
-alias gcm="git commit -m"
-alias gco="git checkout"
-alias gcob="git checkout -b"
-alias gcs="git commit -S -m"
-alias gd="git difftool"
-alias gdc="git difftool --cached"
-alias gf="git fetch"
-alias gg="git graph"
-alias ggg="git graphgpg"
-alias gm="git merge"
-alias gp="git push"
-alias gpr="gh pr create"
-alias gr="git rebase -i"
-alias gs="git status -sb"
-alias gt="git tag"
-alias gu="git reset @ --" #think git unstage
-alias gx="git reset --hard @"
+if (( $+commands[git] )); then
+	alias ga="git add"
+	alias gb="git branch"
+	alias gc="git clone"
+	alias gcm="git commit -m"
+	alias gco="git checkout"
+	alias gcob="git checkout -b"
+	alias gcs="git commit -S -m"
+	alias gd="git difftool"
+	alias gdc="git difftool --cached"
+	alias gf="git fetch"
+	alias gg="git graph"
+	alias ggg="git graphgpg"
+	alias gm="git merge"
+	alias gp="git push"
+	alias gpr="gh pr create"
+	alias gr="git rebase -i"
+	alias gs="git status -sb"
+	alias gt="git tag"
+	alias gu="git reset @ --" #think git unstage
+	alias gx="git reset --hard @"
+fi
 
 #overrides
 alias sudo="sudo "  # expand aliases with sudo
@@ -69,7 +78,7 @@ alias systemctl="sudo systemctl"
 alias proxychains="proxychains -q"
 alias ag="ag --color --color-line-number '0;35' --color-match '46;30' --color-path '4;36'"
 alias aga="ag --hidden --color --color-line-number '0;35' --color-match '46;30' --color-path '4;36'"
-alias tree='tree -CFa -I "CVS|*.*.package|.svn|.git|.hg|node_modules|bower_components" --dirsfirst'
+alias tree='tree -CAFa -I "CVS|*.*.package|.svn|.git|.hg|node_modules|bower_components" --dirsfirst'
 
 #curl
 alias curlh="curl -sILX GET"
@@ -88,3 +97,10 @@ alias xyzzy="echo nothing happens"
 alias lol="base64 </dev/urandom | lolcat"
 alias matrix="cmatrix -b"
 alias zen="while :; do bonsai -l -b 2 -c oO0 -t 0.5; sleep 10; done"
+
+#mac
+if [[ "$OSTYPE" == darwin* ]]; then
+  alias unquarantine='xattr -r -d com.apple.quarantine'
+fi
+
+alias claudework="CLAUDE_CONFIG_DIR=$HOME/.claudework claude"
